STYLE: Moving things around and cleanup
