+++ src/components/search/concept-modal/tabs/cdes/related-studies/index.js
@@ -1,0 +1,1 @@
+export * from './related-studies-list'
